incorporating juliens unified testing scheme with new structure
